Day 11 Part 2

+++ AdventOfCode2023/Schema/Day11/GalaxyPosition.cs
@@ -1,9 +1,9 @@
 ﻿namespace AdventOfCode2023.Schema.Day11;
 
-public record struct GalaxyPosition(int Row, int Column)
+public record struct GalaxyPosition(long Row, long Column)
 {
-    public int Row = Row;
-    public int Column = Column;
+    public long Row = Row;
+    public long Column = Column;
 
     public static GalaxyPosition operator +(in GalaxyPosition lhs, in GalaxyPosition rhs) =>
         new(lhs.Row + rhs.Row,
